refactor(@maz-ui/themes): split color-parser into runtime-minimal and conversions modules
Extract the heavy OKLCh/HSL/RGB conversion math (parseColor, parseColorAsOklch,
colorToHex, formatAsHSL, formatAsOklch) into a new `color-conversions.ts`
module. `color-parser.ts` keeps only the tiny runtime predicates
(`isCompleteCSSColor`, `normalizeColor`) used by the CSS generator on every
render — the heavy math no longer ships in that path.

The runtime now loads ~700b of color-parser instead of ~5.7KB; the conversion
helpers are dynamically reachable for consumers that need them (e.g. the docs'
ColorPicker, third-party tooling).

Public surface preserved: same exports, both reachable via
`@maz-ui/themes/utils` thanks to the wildcard subpath.
