+++ packages/themes/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './color-conversions'
 export * from './color-parser'
 export * from './color-scheme-meta'
 export * from './cookie-storage'
